cleanup dups
