+++ lib/pool.js
@@ -7,7 +7,7 @@ const Pool = function(opts, sandCastleCreationOpts) {
     numberOfInstances: 1
   }, opts);
 
-  if(this.numberOfInstances < 1) {
+  if (this.numberOfInstances < 1) {
     throw "Can't create a pool with zero instances";
   }
 
@@ -25,22 +25,22 @@ const Pool = function(opts, sandCastleCreationOpts) {
     this.createMmap(0);
   }
 
-  let sandCastleOptions = {};
-  for(let i = 0; i < this.numberOfInstances; ++i) {
+  const sandCastleOptions = {};
+  for (let i = 0; i < this.numberOfInstances; ++i) {
     Object.assign( sandCastleOptions,
-              sandCastleCreationOpts, {
-                socket: `/tmp/sandcastle_${i.toString()}.sock`,
-                sharedFileID: this.ufid,
-                sharedObject: this.mmapObject
-              }
-            );
+      sandCastleCreationOpts, {
+        socket: `/tmp/sandcastle_${i.toString()}.sock`,
+        sharedFileID: this.ufid,
+        sharedObject: this.mmapObject
+      }
+    );
 
     this.sandcastles.push({
       castle: new SandCastle(sandCastleOptions),
       running: false
     });
   }
-}
+};
 
 Pool.prototype.createMmap = function(memoryLimit) {
   // create a unique file id in case of multiple or crash
@@ -51,8 +51,8 @@ Pool.prototype.createMmap = function(memoryLimit) {
   // setup with a portion of memory (we only want a ~10th of the max)
   let mmapMemory = memoryLimit * 100;
   if (mmapMemory === 0) { mmapMemory = 5120; } //KB
-  const mmapObjects = (this.numberOfInstances*2)+1 // x2 + 1 for two scripts and the type
-  this.mmapObject = new MMAP.Create(tempMemFile, mmapMemory, this.numberOfInstances*2);
+  const mmapObjects = (this.numberOfInstances*2)+1; // x2 + 1 for two scripts and the type
+  this.mmapObject = new MMAP.Create(tempMemFile, mmapMemory, mmapObjects);
 };
 
 Pool.prototype.kill = function() {
