+++ lib/sandcastle.js
@@ -53,7 +53,7 @@ const SandCastle = function(opts) {
   process.on('exit', this.onProcessExit);
 };
 
-SandCastle.prototype.createMmap = function () {
+SandCastle.prototype.createMmap = function() {
   // create a unique file id in case of multiple or crash
   const pid = process.pid ? process.pid.toString(36) : '';
   const time = Date.now().toString(36);
